Install developer certificate and fix dotnet run not cleaning up (#364)
* Installs an ASP.NET Core developer certificate as a Cake task, so that we can run dotnet run against projects created from a template. This fails on Mac because Azure Pipelines because we don't have admin permissions and fails on Ubuntu because the .NET API to install certificates doesn't exist.
* Enables running dotnet run based unit tests on Windows but not Linux and Mac.
* Fix dotnet run command that was not cleaning up child processes when we were done with it and wanted to shut down the app.

+++ build.cake
@@ -1,3 +1,4 @@
+using System.Diagnostics;
 using System.Security.Cryptography.X509Certificates;
 
 var target = Argument("Target", "Default");
@@ -22,6 +23,7 @@ var artifactsDirectory = Directory("./Artifacts");
 var templatePackProject = Directory("./Source/*.csproj");
 var versionSuffix = string.IsNullOrEmpty(preReleaseSuffix) ? null : preReleaseSuffix + "-" + buildNumber.ToString("D4");
 var isRunningOnCI = TFBuild.IsRunningOnAzurePipelinesHosted || AppVeyor.IsRunningOnAppVeyor;
+var isDotnetRunEnabled = !isRunningOnCI || (isRunningOnCI && IsRunningOnWindows());
 
 Task("Clean")
     .Does(() =>
@@ -52,31 +54,36 @@ Task("Restore")
     });
 
 Task("InstallDeveloperCertificate")
-    .WithCriteria(x => TFBuild.IsRunningOnAzurePipelinesHosted || AppVeyor.IsRunningOnAppVeyor)
     .Does(() =>
     {
-        var certificateFilePath = System.IO.Path.ChangeExtension(System.IO.Path.GetTempFileName(), ".pfx");
-        StartProcess(
-            "dotnet",
-            new ProcessArgumentBuilder()
-                .Append("dev-certs")
-                .Append("https")
-                .AppendSwitch("--export-path", certificateFilePath));
-        Information($"Dotnet Developer Certificate saved");
-
-        var certificate = new X509Certificate2(certificateFilePath);
-        using (var store = new X509Store(StoreName.Root, StoreLocation.LocalMachine))
+        if (isDotnetRunEnabled)
         {
-            store.Open(OpenFlags.ReadWrite);
-            store.Add(certificate);
-        }
-        Information($"Dotnet Developer Certificate installed to local machine");
+            var certificateFilePath = System.IO.Path.ChangeExtension(System.IO.Path.GetTempFileName(), ".pfx");
+            StartProcess(
+                "dotnet",
+                new ProcessArgumentBuilder()
+                    .Append("dev-certs")
+                    .Append("https")
+                    .AppendSwitch("--export-path", certificateFilePath));
+            Information($"Dotnet Developer Certificate saved");
 
-        System.IO.File.Delete(certificateFilePath);
+            var certificate = new X509Certificate2(certificateFilePath);
+            using (var store = new X509Store(StoreName.Root, StoreLocation.LocalMachine))
+            {
+                store.Open(OpenFlags.ReadWrite);
+                store.Add(certificate);
+            }
+            Information($"Dotnet Developer Certificate installed to local machine");
+
+            System.IO.File.Delete(certificateFilePath);
+        }
+        else
+        {
+            Information("This CI server does not support installing certificates");
+        }
     });
 
 Task("Test")
-    .IsDependentOn("InstallDeveloperCertificate")
     .Does(() =>
     {
         foreach(var project in GetFiles("./Tests/**/*.csproj"))
@@ -86,12 +93,13 @@ Task("Test")
                 new DotNetCoreTestSettings()
                 {
                     Configuration = configuration,
-                    Filter = isRunningOnCI ? "IsUsingDotnetRun=false" : null,
+                    Filter = isDotnetRunEnabled ? null : "IsUsingDotnetRun=false",
                     Logger = $"trx;LogFileName={project.GetFilenameWithoutExtension()}.trx",
                     NoBuild = true,
                     NoRestore = true,
                     ResultsDirectory = artifactsDirectory
                 });
+            Information($"Completed {project.GetFilenameWithoutExtension()} tests");
         }
     });
 
@@ -116,18 +124,6 @@ Task("Default")
     .IsDependentOn("Pack");
 
 RunTarget(target);
-
-Teardown(context =>
-{
-    // Appveyor is failing to exit the cake script.
-    if (AppVeyor.IsRunningOnAppVeyor)
-    {
-        foreach (var process in System.Diagnostics.Process.GetProcessesByName("dotnet"))
-        {
-            process.Kill();
-        }
-    }
-});
 
 public void StartProcess(string processName, ProcessArgumentBuilder builder)
 {
